data: using Germany's postal code db
source: https://www.suche-postleitzahl.org/downloads?download=plz-gebiete.shp.zip

the file is a cleaned up version of `plz_einwohner` - contains the last
three columns, minus decimal places for qkm and no dots for lat/long.

this is (hopefully) similar to examples in Wikipedia, which might give a
good idea on the distribution of data.